[MODLOGSAML-170] - Refactoring
