+++ src/main/java/org/folio/util/UserFields.java
@@ -1,0 +1,15 @@
+package org.folio.util;
+
+public class UserFields {
+
+  public static final String ID = "id";
+  public static final String EMAIL = "email";
+  public static final String USER_ID = "userId";
+  public static final String BARCODE = "barcode";
+  public static final String USERNAME = "username";
+  public static final String TENANT_ID = "tenantId";
+  public static final String USER_TENANTS = "userTenants";
+  public static final String TOTAL_RECORDS = "totalRecords";
+  public static final String PERSONAL_EMAIL = "personal.email";
+  public static final String EXTERNAL_SYSTEM_ID = "externalSystemId";
+}
